Minor fixes and comments in README
I installed the pact_broker (thanks for that) on a local Windows system, and fixed a few minor issues in the README. I also added a hint on how Windows users get Rails/Ruby. Since Windows users aren't generally very familiar with Ruby, I believe this makes sense. You may like these edits or not, they are not spectacular, but I did not want to keep the findings for myself. Thanks for your work with pact!