feat: add configuration option for check_for_potential_duplicate_pacticipant_names
This allows the checking for potential duplicate names (eg. a service being called "Foo" in
one pact and "Foo Service" or "foo" or "Foos" in another) to be turned off when publishing a pact.

Turning this check off will likely result in a mess of data, and result in the end of the world as we know it.

Just for @uglyog